perf(web/myspace): upgrade myspace
